+++ lib/views/about.dart
@@ -201,9 +201,6 @@ class Avatar extends StatelessWidget {
           Text(contributor.name, style: context.textTheme.bodySmall),
         ],
       ),
-      // onTap: () {
-      //   globalState.openUrl(contributor.link);
-      // },
     );
   }
 }